fix bug: the animation of the + icon not working on web
